+++ src/quaternion.cpp
@@ -32,15 +32,15 @@ either expressed or implied, of the FreeBSD Project.
 
 using namespace sv;
 
-Quaternion::Quaternion(const double angle, const cv::Vec3f &axis):internal_quaternion_(angle,axis[0],axis[1],axis[2]){}
+Quaternion::Quaternion(const double angle, const cv::Vec3d &axis):internal_quaternion_(angle,axis[0],axis[1],axis[2]){}
 
-Quaternion Quaternion::FromVectorToVector(const cv::Vec3f &from, const cv::Vec3f to){
+Quaternion Quaternion::FromVectorToVector(const cv::Vec3d &from, const cv::Vec3d to){
   
-  cv::Vec3f from_n,to_n;
+  cv::Vec3d from_n,to_n;
   cv::normalize(from,from_n);
   cv::normalize(to,to_n);
       
-  float d = from_n.dot(to_n);
+  double d = from_n.dot(to_n);
 
   if(d >= 1.0){
     return boost::math::quaternion<double>();
@@ -48,12 +48,12 @@ Quaternion Quaternion::FromVectorToVector(const cv::Vec3f &from, const cv::Vec3f
 
   //check if d \approx = 0
 
-  float s = (float)sqrt( (1+d)*2 );
-  float inv_s = 1.0f/s;
+  double s = (double)sqrt( (1+d)*2 );
+  double inv_s = 1.0f/s;
 
-  cv::Vec3f axis = from_n.cross(to_n);
+  cv::Vec3d axis = from_n.cross(to_n);
 
-  Quaternion q( s*0.5f, cv::Vec3f(axis[0]*inv_s, axis[1]*inv_s, axis[2]*inv_s ));
+  Quaternion q( s*0.5f, cv::Vec3d(axis[0]*inv_s, axis[1]*inv_s, axis[2]*inv_s ));
 
   return q.Normalize();
 
@@ -75,12 +75,12 @@ Quaternion Quaternion::Normalize() const {
                                          internal_quaternion_.R_component_4()/mag);
 }
 
-cv::Vec3f Quaternion::RotateVector(const cv::Vec3f &to_rotate) const {
+cv::Vec3d Quaternion::RotateVector(const cv::Vec3d &to_rotate) const {
    
   const boost::math::quaternion<double> vec_quat(0,to_rotate[0],to_rotate[1],to_rotate[2]);
 
   boost::math::quaternion<double> rotated = (internal_quaternion_ * vec_quat) * boost::math::conj<double>(internal_quaternion_);
 
-  return cv::Vec3f((float)rotated.R_component_2(),(float)rotated.R_component_3(),(float)rotated.R_component_3());
+  return cv::Vec3d((double)rotated.R_component_2(),(double)rotated.R_component_3(),(double)rotated.R_component_3());
 
 }
